+++ src/votes/types/user-vote-response.type.ts
@@ -1,0 +1,5 @@
+import { VoteType } from '@prisma/client';
+
+export type UserVoteResponse = {
+  type: VoteType | null;
+};